feat: Replaced hacky `hr` effect with css
Replaced the usage of the `hr` element
with the a css class `about-myself`. This
will make the page more accessible as the
`hr` tag had no semantics with regard to
the document content. The `hr` tag was
added to move the first in the `main-area`
lower than the end of the `header`. We now
do this via css.